# spec/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api Documentation",
    "version" : "1.0",
    "title" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/users/create" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "create",
        "operationId" : "createUsingPOST_4",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "userDTO",
          "description" : "userDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/users/delete/{id}" : {
      "delete" : {
        "tags" : [ "user-controller" ],
        "summary" : "delete",
        "operationId" : "deleteUsingDELETE_3",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          },
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
    "/api/users/email/{email}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getUserByEmail",
        "operationId" : "getUserByEmailUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "description" : "email",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/users/id/{id}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getById",
        "operationId" : "getByIdUsingGET_2",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/users/list" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "list",
        "operationId" : "listUsingGET_5",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/users/save/{id}" : {
      "put" : {
        "tags" : [ "user-controller" ],
        "summary" : "save",
        "operationId" : "saveUsingPUT_5",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "userDTO",
          "description" : "userDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/v1/users/create" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "create",
        "operationId" : "createUsingPOST_5",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "userDTO",
          "description" : "userDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/users/delete/{id}" : {
      "delete" : {
        "tags" : [ "user-controller" ],
        "summary" : "delete",
        "operationId" : "deleteUsingDELETE_4",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          },
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
    "/api/v1/users/email/{email}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getUserByEmail",
        "operationId" : "getUserByEmailUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "description" : "email",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/users/id/{id}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getById",
        "operationId" : "getByIdUsingGET_3",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/users/list" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "list",
        "operationId" : "listUsingGET_6",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/users/save/{id}" : {
      "put" : {
        "tags" : [ "user-controller" ],
        "summary" : "save",
        "operationId" : "saveUsingPUT_6",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "userDTO",
          "description" : "userDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "UserDTO" : {
      "type" : "object",
      "properties" : {
        "adId" : {
          "type" : "string"
        },
        "associatedPlace" : {
          "type" : "string"
        },
        "department" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "emailRegionalManager" : {
          "type" : "string"
        },
        "emailSalesManager" : {
          "type" : "string"
        },
        "fullName" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "jobTitle" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "orgName" : {
          "type" : "string"
        },
        "orgNr" : {
          "type" : "string"
        },
        "overallPowerOfAttorney" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "powerOfAttorneyFA" : {
          "type" : "string"
        },
        "powerOfAttorneyOA" : {
          "type" : "string"
        },
        "regionName" : {
          "type" : "string"
        },
        "regionalManagersPowerOfAttorney" : {
          "type" : "string"
        },
        "resourceNr" : {
          "type" : "string"
        },
        "salesRoleId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "sureName" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "usernameAlias" : {
          "type" : "string"
        }
      },
      "title" : "UserDTO"
    },
...
  }
}